hr_bot/requirements.txt

botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-ai>=4.14.0
botframework-connector>=4.14.0
aiohttp>=3.8.1
azure-identity>=1.7.0
azure-cognitiveservices-language-luis>=0.7.0
openai>=0.27.0
python-dotenv>=0.19.0
requests>=2.28.0
reportlab>=3.6.8
fpdf>=1.7.2
gunicorn>=20.1.0
autopep8>=1.6.0
pylint>=2.12.0